requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
backports.zoneinfo==0.2.1; python_version < '3.9'
django==4.2.16; python_version >= '3.8'
djangorestframework==3.15.2; python_version >= '3.8'
drf-yasg==1.21.7; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
packaging==24.1; python_version >= '3.8'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
sqlparse==0.5.1; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.11'
uritemplate==4.1.1; python_version >= '3.6'
